Conversão de tipos do campo quantidade

+++ OlimpiadaCompras/OlimpiadaCompras/Models/ProdutoPedidoOrcamento.cs
@@ -9,11 +9,11 @@ namespace OlimpiadaCompras.Models
     public class ProdutoPedidoOrcamento
     {
         public long Id { get; set; }
-        public double valor { get; set; }
-        public int Quantidade { get; set; }
-        public double Ipi { get; set; }
-        public double Icms { get; set; }
-        public double Desconto { get; set; }
+        public decimal valor { get; set; }
+        public decimal Quantidade { get; set; }
+        public decimal Ipi { get; set; }
+        public decimal Icms { get; set; }
+        public decimal Desconto { get; set; }
         public Orcamento Orcamento { get; set; }
         public ProdutoSolicitacao ProdutoSolicitacao { get; set; }
         public long ProdutoSolicitacoesId { get; set; }
